Fix usage of markdown inside HTML blocks
Interaction between markdown and HTML blocks is currently complex
and has many edge cases, especially related to whitespace and blank lines.
See the examples in the spec: https://spec.commonmark.org/0.29/#html-blocks

This was the cause of some unprocessed formatting (mostly code fenced
with backticks) inside some of the `<aside>` blocks, as can be seen
in the `_cards.md`, `_entities.md` and `_transactions.md` pages.

As a solution (and standardizing on a practice that was already present
in various pages of this project) all formatting inside `<aside>` blocks
is converted to HTML.

This way the content becomes robust to such flaky interactions,
and the existing cases are fixed.